requirements.txt
Flask==3.0.0
numpy==1.26.2
openai==1.3.7
pandas==2.1.3
python-dotenv==1.0.0
